fix: :bug: fixed security hotspots by sonarqube CI

+++ backend/src/main/java/com/DraftLeague/services/MarketScheduler.java
@@ -38,7 +38,7 @@ public class MarketScheduler {
 
             for (League league : leagues) {
                 try {
-                    marketService.refreshMarket(league.getId().intValue());
+                    marketService.refreshMarket(league.getId());
                 } catch (Exception e) {
                     logger.error("Error al refrescar mercado para la liga {}: {}", league.getId(), e.getMessage(), e);
                 }
